requirements.txt
alembic
factory_boy
mysqlclient
# mysql-connector-python-rf
sqlalchemy
sqlalchemy_utils
